Handle ignored /src dir.

+++ bin/slideshow.py
@@ -87,8 +87,9 @@ def save_slide():
         for path in source_paths():
             dst_dir = Path("slides", current)
             dst_dir.mkdir(exist_ok=True)
-            print("Saving {} to {}".format(path, dst_dir))
-            shutil.copy2(str(path), str(dst_dir))
+            if path.exists():
+                print("Saving {} to {}".format(path, dst_dir))
+                shutil.copy2(str(path), str(dst_dir))
         print("Saved slide {}".format(current))
     else:
         print("Read-only mode, not saving")
@@ -96,6 +97,7 @@ def save_slide():
 
 def switch_to_slide(slide):
     for path in source_paths():
+        path.parent.mkdir(parents=True, exist_ok=True)
         src_path = Path("slides", slide, path.name)
         if src_path.exists():
             print("Copying {} to {}".format(src_path, path))
